remove text asset functionality, move color extractor to separate file
