+++ .gitignore
@@ -4,3 +4,5 @@ skottie-player.d.ts
 skottie-player.d.ts.map
 skottie-player.js
 skottie-player.js.map
+color-extractor.d.ts
+color-extractor.d.ts.map
