+++ src/skottie-player.ts
@@ -7,7 +7,8 @@ import JSONEditor, {JSONEditorOptions} from 'jsoneditor';
 
 const rgbToHsv = require('rgb-hsv');
 const hsvToRgb = require('hsv-rgb');
-const { rgb2lab, lab2rgb, deltaE } = require('rgb-lab');
+const { rgb2lab, deltaE } = require('rgb-lab');
+const { getDominantColors } = require('./color-extractor');
 const SkottieKitInit = require('skottiekit-wasm/bin/skottiekit.js');
 const loadKit = SkottieKitInit({
   locateFile: (file: string) => '/node_modules/skottiekit-wasm/bin/'+file,
@@ -325,6 +326,7 @@ export class SkottiePlayer extends LitElement {
         </div>
       `;
     } else if (layer.ty === 4) {
+      // layer is shape layer
       return html`
         <br>
         ${(layer.shapes as any[]).map(shape => shape.ty === "fl" || shape.ty === "gr" ?
@@ -360,36 +362,7 @@ export class SkottiePlayer extends LitElement {
     }
   }
 
-  /**
-   * Renders text inputs to change text assets in the animation.
-   */
-  renderTextAssets(): TemplateResult | string {
-    if (this.content.assets) {
-      return html`
-        ${this.content.assets.map(asset => asset.layers !== undefined ?
-          html`
-          ${(asset.layers as Layer[]).map(layer => layer.ty === 5 ?
-            layer.t.d?.k.map((el: { s: { t: any; }; }) => el.s.t !== undefined ?
-              html`
-              ${layer.nm}: 
-              <input
-              id="${asset.id + ' ' + layer.nm} Input"
-              type="text"
-              value="${el.s.t}"
-              @change="${this.updateJson}">
-              <br>
-              `:
-              ""  
-            ):
-            ""
-          )}
-          `:
-          ""
-        )}
-      `;
-    }
-    return '';
-  }
+  
 
   render(): TemplateResult {
     return html`
@@ -443,8 +416,6 @@ export class SkottiePlayer extends LitElement {
             value="${this.content.fr}"
             @change="${this.updateJson}">
             </div>
-
-            ${this.renderTextAssets()}
 
             ${(this.content.layers as Layer[]).map(layer => 
               html`
@@ -625,11 +596,17 @@ export class SkottiePlayer extends LitElement {
       canvasElement.setAttribute('height', this.height.toString());
       const context = canvasElement.getContext('2d');
       if (context) {
-        context.drawImage(img, 0, 0, img.width * this.height / img.height, this.height);
-        const myData = context.getImageData(0, 0, img.width * this.height / img.height, this.height);
+        context.drawImage(img, 0, 0, 
+          img.width * this.height / img.height, this.height);
+        const myData = context.getImageData(0, 0,
+          img.width * this.height / img.height, this.height);
         
-        const k = 5;
-        let [centers, centerCounts] = this.getDominantColors(myData, k);
+        const numClusters = 5;
+        const colorThreshold = 30;
+        const white = '000000000';
+        const black = '255255255'
+        let [centers, centerCounts] = getDominantColors(myData, numClusters, 
+          [white, black], colorThreshold);
         let [index, max] = [0, 0];
         for (let i = 0; i < centerCounts.length; i++) {
           if (centerCounts[i] > max) {
@@ -648,334 +625,6 @@ export class SkottiePlayer extends LitElement {
       URL.revokeObjectURL(url);
     }
     img.src = url;
-  }
-
-  /**
-   * Takes in an ImageData object and finds the k most dominant colors from
-   * the image data. Also returns the number of pixels assoiated with each 
-   * color's cluster.
-   * @param myData  ImageData generated from a HTML Canvas element
-   */
-  private getDominantColors(myData: ImageData, k: number): Array<any> {
-    let dataMap, keyArray, total;
-    [dataMap, keyArray, total] = this.getDataMap(myData)
-
-    let centerCandidates = this.getCenterCandidates(dataMap, keyArray, total)
-
-    let centers = new Array<string>();
-    let centerCounts, centerToColor;
-    while (centers.length === 0) {
-      [centers, centerCounts, centerToColor] = this.kMeans(k, dataMap, centerCandidates);
-      [centers, centerCounts] = this.filterWhiteBlack(centers, centerCounts, centerToColor);
-    }
-
-    return [centers, centerCounts];
-  }
-
-  /**
-   * Gets center candidates for performing k means on the data in dataMap.
-   * Eliminates candidates if the count for a certain color is too low.
-   * 
-   * @param dataMap   Map from color to count of color in image.
-   * @param keyArray  Array containing keys of dataMap in arbitrary order.
-   * @param total     Total number of pixels from image used to make dataMap.
-   */
-  private getCenterCandidates(dataMap: Map<string, number>, 
-    keyArray: Array<string>, total: number) {
-    let centerCandidates = [];
-    for (let i = 0; i < keyArray.length; i++) {
-      const count = dataMap.get(keyArray[i])
-      if (count && total / keyArray.length / 2 < count){
-        centerCandidates.push(keyArray[i]);
-      }
-    }
-    return centerCandidates
-  }
-
-  /**
-   * Filters out cluster centers that are too similar to white or black.
-   * Also converts center color to nearest color found in the image.
-   * Returns array of valid centers and center counts for those centers.
-   * 
-   * @param centers         Array of colors that are centers.
-   * @param centerCounts    Map from center to count of number of pixels.
-   * @param centerToColor   Map from center to colors in center's cluster.
-   */
-  private filterWhiteBlack(centers: Array<string>,
-    centerCounts: Array<number>,
-    centerToColor: Map<string, Array<string>>): Array<any> {
-    const black = '000000000';
-    const white = '255255255';
-    const threshold = 30;
-    for (let i = centers.length - 1; i >= 0; i--) {
-      if (this.colorDistance(white, centers[i]) > threshold
-      && this.colorDistance(black, centers[i]) > threshold) {
-        centers[i] = this.getImageColors(centerToColor, centers[i]);
-      } else {
-        centers.splice(i, 1);
-        centerCounts.splice(i, 1);
-      }
-    }
-    return [centers, centerCounts]
-  }
-
-  /**
-   * Converts an ImageData object into a map from colors to counts in the 
-   * image. Also returns an array of all colors and the total number of pixels
-   * in the image. 
-   */
-  private getDataMap(myData: ImageData): Array<any> {
-    let dataMap = new Map<string, number>();
-    let keyArray: Array<string> = [];
-    let total = 0;
-    for (let i = 0; i < myData.data.length; i += 4) {
-      if (myData.data[i + 3] > 0) {
-        let r = myData.data[i].toString();
-        while (r.length < 3) {
-          r = '0' + r;
-        }
-        let g = myData.data[i + 1].toString();
-        while (g.length < 3) {
-          g = '0' + g;
-        }
-        let b = myData.data[i + 2].toString();
-        while (b.length < 3) {
-          b = '0' + b;
-        }
-
-        const key = r + g + b;
-        const count = dataMap.get(key);
-        if (count) {
-          dataMap.set(key, count + 1);
-        } else {
-          dataMap.set(key, 1);
-          keyArray.push(key);
-        }
-        total++;
-      }
-    }
-    return [dataMap, keyArray, total]
-  }
-
-  /**
-   * Converts a center of a cluster generated from k means to the closest color
-   * found in the image represented as a hex string.
-   * @param centerToColor  Map from center color to all colors in cluster
-   * @param color          Color of center to convert.
-   */
-  private getImageColors(centerToColor: any, color: string): string {
-    const colorArray = centerToColor.get(color)
-    let minDistColor = '';
-    if (colorArray) {
-      let minDist = deltaE(rgb2lab([0, 0, 0]), rgb2lab([255, 255, 255]));
-      for (let i = 0; i < colorArray.length; i++) {
-        const newDist = this.colorDistance(colorArray[i], color);
-        if (newDist < minDist) {
-          minDist = newDist;
-          minDistColor = colorArray[i];
-        }
-      }
-      minDistColor = this.rgbaToHex([parseInt(minDistColor.slice(0,3)) / 255,
-        parseInt(minDistColor.slice(3,6)) / 255,
-        parseInt(minDistColor.slice(6)) / 255]);
-    }
-    return minDistColor;
-  }
-
-  /**
-   * Performs k-means with points in color space. Points are represented in a
-   * color to count map. Returns the centers of clusters, the number of points
-   * in each cluster, and a map from centers to all colors in its cluster.
-   */
-  private kMeans(k: number, data: Map<string, number>, 
-    keys: Array<string>): Array<any> {
-    let centers = this.initializeCenters(k, keys);
-    let centerCounts, centerToColor;
-    let oldCenters = centers.slice();
-    while (true) {
-      [centers, centerCounts, centerToColor] = this.updateCenters(centers, data, keys);
-      if (this.arraysEqual(centers, oldCenters)) {
-        break;
-      }
-      oldCenters = centers.slice();
-    }
-    return [centers, centerCounts, centerToColor];
-  }
-
-  /**
-   * Randomly initializes centers for k-means algorithm. Chooses k random 
-   * colors from an array of possible colors to be intial centers.
-   */
-  private initializeCenters(k: number, keys: Array<string>):
-    Array<string> {
-    if (k >= keys.length) {
-      return keys;
-    }
-    let centers: Array<string> = [];
-    let usedIndices = new Set<number>();
-    while (centers.length < k) {
-      let index = Math.floor(Math.random() * keys.length);
-      while (usedIndices.has(index)) {
-        index = Math.floor(Math.random() * keys.length);
-      }
-      centers.push(keys[index]);
-      usedIndices.add(index);
-    }
-    return centers;
-  }
-
-  /**
-   * Updates centers for k means. First reassigns pixels to new clusters,
-   * then recalculates cluster centers based on reassignments. Returns
-   * new centers, counts for new centers, and a map from each center to
-   * an array of the colors assigned to that center's cluster.
-   * 
-   * @param centers   Array of old centers.
-   * @param data      Map from color to count of color in image.
-   * @param keys      Array of all colors in image.
-   */
-  private updateCenters(centers: Array<string>, data: Map<string, number>,
-    keys: Array<string>): Array<any> {
-
-    // calculate new clusters
-    const pointToCenter = this.getNewClusters(keys, centers);
-
-    // calculate new centers
-    const [centerCounts, centerSums] = 
-      this.getCenterStats(centers.length, keys, data, pointToCenter);
-    const newCenters = this.getNewCenters(centerCounts, centerSums);
-
-    // generate map from center to all colors in cluster
-    let centerToColor: Map<string, Array<string>> = 
-      new Map<string, Array<string>>();
-    for (let i = 0; i < newCenters.length; i++) {
-      centerToColor.set(newCenters[i], []);
-    }
-    for (let i = 0; i < keys.length; i++) {
-      centerToColor.get(newCenters[pointToCenter[i]])?.push(keys[i]);
-    }
-
-    return [newCenters, centerCounts, centerToColor];
-  }
-
-  /**
-   * Gets new cluster assignments based on colors and centers.
-   * 
-   * @param colors   Array of all colors.
-   * @param centers  Array of current centers.
-   */
-  private getNewClusters(colors: Array<string>, 
-    centers: Array<string>): Array<number> {
-    let pointToCenter: Array<number> = [];
-    for (let i = 0; i < colors.length; i++) {
-      let minDist = 100;
-      let currCenter = 0;
-      for (let j = 0; j < centers.length; j++ ) {
-        const newDist = this.colorDistance(colors[i], centers[j]);
-        if (newDist < minDist) {
-          minDist = newDist;
-          currCenter = j;
-        }
-      }
-      pointToCenter.push(currCenter);
-    }
-    return pointToCenter
-  }
-
-  /**
-   * Gets size of each cluster as well as the sum of all pixel colors of each 
-   * cluster represented as an array of rgb array.
-   * 
-   * @param k              Number of centers
-   * @param keys           Array of all colors in image.
-   * @param data           Map from color to count of color in image.
-   * @param pointToCenter  Array that indicates which center which color 
-   *                       coresponds to. Same length as keys.
-   */
-  private getCenterStats(k: number, keys: Array<string>, 
-    data: Map<string, number>, pointToCenter: Array<number>): Array<any> {
-    let centerCounts: Array<number> = [];
-    let centerSums: Array<Array<number>> = [];
-    for (let i = 0 ; i < k; i++) {
-      centerCounts.push(0);
-      centerSums.push([]);
-      for (let j = 0; j < 3; j++) {
-        centerSums[i].push(0);
-      }
-    }
-
-    for (let i = 0; i < keys.length; i++) {
-      const multiplier = data.get(keys[i]);
-      if (multiplier) {
-        centerSums[pointToCenter[i]][0] += parseInt(keys[i].slice(0,3)) * multiplier;
-        centerSums[pointToCenter[i]][1] += parseInt(keys[i].slice(3,6)) * multiplier;
-        centerSums[pointToCenter[i]][2] += parseInt(keys[i].slice(6)) * multiplier;
-        centerCounts[pointToCenter[i]] += multiplier;
-      }
-    }
-
-    for (let i = 0; i < centerCounts.length; i++) {
-      for (let j = 0; j < 3; j++) {
-        centerSums[i][j] = Math.round(centerSums[i][j] / centerCounts[i]);
-      }
-    }
-    return [centerCounts, centerSums];
-  }
-
-  /**
-   * Gets new centers based on count of pixels per center and sum of pixel 
-   * colors per center. Returns array of new centers represented as hex 
-   * strings.
-   */
-  private getNewCenters(centerCounts: Array<number>, 
-    centerSums: Array<Array<number>>): Array<string> {
-    let newCenters: Array<string> = [];
-    for (let i = 0; i < centerCounts.length; i++) {
-      let r = centerSums[i][0].toString();
-      while (r.length < 3) {
-        r = '0' + r;
-      }
-      let g = centerSums[i][1].toString();
-      while (g.length < 3) {
-        g = '0' + g;
-      }
-      let b = centerSums[i][2].toString();
-      while (b.length < 3) {
-        b = '0' + b;
-      }
-      newCenters.push(r + g + b);
-    }
-    return newCenters;
-  }
-
-  /**
-   * Checks if two string arrays have the same elements in every spot. 
-   */
-  private arraysEqual(array1: Array<string>, array2: Array<string>): boolean {
-    if (array1 === array2) {
-      return true;
-    } else if (!array1 || !array2 || array1.length !== array2.length) {
-      return false;
-    }
-
-    for (let i = 0; i < array1.length; i++) {
-      if (array1[i] !== array2[i]) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  /**
-   * Gets the lab color distance between two RGB color strings.
-   */
-  private colorDistance(color1: string, color2: string): number {
-    const labColor1 = rgb2lab([parseInt(color1.slice(0,3)), 
-      parseInt(color1.slice(3,6)), parseInt(color1.slice(6))]);
-    const labColor2 = rgb2lab([parseInt(color2.slice(0,3)),
-      parseInt(color2.slice(3,6)), parseInt(color2.slice(6))]);
-    return deltaE(labColor1, labColor2);
   }
 
   /**
@@ -1109,7 +758,6 @@ export class SkottiePlayer extends LitElement {
         this.setContent((this.jsonEditor as JSONEditor).get());
       } else {
         this.updateJsonFramerate();
-        this.updateJsonTextAssets();
         this.updateJsonLayers();
       }
       this.fileString = JSON.stringify(this.content);
@@ -1223,39 +871,6 @@ export class SkottiePlayer extends LitElement {
     const framerateInput = (this.shadowRoot?.getElementById('newframerate') as HTMLInputElement);
     if (framerateInput.value) {
       this.content.fr = parseFloat(framerateInput.value);
-    }
-  }
-
-  /**
-   * Updates the text assets in the content object.
-   */
-  private updateJsonTextAssets(): void {
-    if (this.content.assets) {
-      for (let i = 0; i < this.content.assets.length; i++) {
-        if (this.content.assets[i].layers) {
-          for (let j = 0; j < this.content.assets[i].layers.length; j++) {
-            this.updateAssetText(i, j);
-          }
-        }
-      }
-    }
-  }
-
-  /**
-   * Updates a specific text asset in the content object.
-   */
-  private updateAssetText(assetIndex: number, layerIndex: number): void {
-    if (this.content.assets && this.content.assets[assetIndex].layers) {
-      const textInput = (this.shadowRoot?.getElementById(this.content.assets[assetIndex].id
-        + ' ' + this.content.assets[assetIndex].layers[layerIndex].nm + ' Input') as HTMLInputElement);
-      if (textInput && this.content.assets[assetIndex].layers[layerIndex].t.d.k) {
-        for (let i = 0; i < this.content.assets[assetIndex].layers[layerIndex].t.d.k.length; i++) {
-          if (this.content.assets[assetIndex].layers[layerIndex].t.d.k[i].s) {
-            this.content.assets[assetIndex].layers[layerIndex].t.d.k[i].s.t = textInput.value;
-            return;
-          }
-        }
-      }
     }
   }
